Case study
This is a case study where we applied our activation functions in a convolutional model to train and test on the MNIST dataset. This dataset consists of images of handwritten digits, and the task is to classify the corresponding number in each image. We demonstrated that the accuracy of our activation function is superior to state-of-the-art activation functions, proving how competitive our NALU2MIG function is.